Tool output moved where it belongs to commons
