+++ src/org/openjdk/asmtools/jasm/Main.java
@@ -23,7 +23,7 @@
 package org.openjdk.asmtools.jasm;
 
 import org.openjdk.asmtools.common.structure.CFVersion;
-import org.openjdk.asmtools.common.structure.ToolInput;
+import org.openjdk.asmtools.common.ToolInput;
 
 import java.io.File;
 import java.io.IOException;
